hotelrateintel-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
bcrypt==4.0.1
pyjwt==2.8.0

# Web scraping
selenium==4.15.0
beautifulsoup4==4.12.2
aiohttp==3.9.0
requests==2.31.0
lxml==4.9.3

# Data processing
pandas==2.1.3
numpy==1.25.2

# Scheduling and background tasks
celery==5.3.4
redis==5.0.1
apscheduler==3.10.4

# Monitoring and logging
sentry-sdk[fastapi]==1.38.0

# Email notifications
sendgrid==6.10.0
jinja2==3.1.2

# Machine learning (for Phase 3)
scikit-learn==1.3.2
tensorflow==2.14.0
prophet==1.1.4

# Data visualization
plotly==5.17.0
matplotlib==3.8.2

# Cloud storage
boto3==1.34.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1 